requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.46.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
executing==2.2.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
huggingface-hub==0.33.2
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
peft==0.16.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchvision==0.22.1+cu126
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
